Update to newest versions of dependencies.

+++ tests/check_shapes/integration/test_jax.py
@@ -176,7 +176,7 @@ def test_check_shapes() -> None:
         test_predictions = model.predict(data.test_features)
         err = data.test_targets - test_predictions
         result = {}
-        result["rmse"] = jnp.sqrt(jnp.mean(err ** 2))
+        result["rmse"] = float(jnp.sqrt(jnp.mean(err ** 2)))
         return result
 
     def main() -> None:
